+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Version 0.0.298 (January 18, 2026)
 - Added guardrails around code execution
-- Removed code validation flag, as it's now determinstic and built-in 
+- Removed code validation flag, as it's now deterministic and built-in 
 
 ## Version 0.0.297 (January 18, 2026)
 - Introducing: Training Mode